requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-environ==0.4.5
django-cors-headers==4.1.0

